Added the ability to search in DiveMenu, so the player can pick up items.
PlayerData:
1. Added methods putIntoHand and putIntoInventory, which accepts either item ID or an ItemEnum.
2. Added method isInventoryFull.
3. Added constant MAX_INVENTORY_SPACE and fields onHand and inventory.

PlayState:
1. Added searchItem clickHandler and applied it to DiveMenu.

DiveMenu:
1. Added ClickHandler setter for searchBox.
2. Added field foundBox, which might get thrown out.

ClickTextBox:
1. Commented out a useless line of code.

Added ItemEnum, which stores information aout each item. Added RandomGenerator, with methods generateGem, which generates a random item ID, and generateBatteryDrain, which determines if the oil rig should lose a battery after a player buys oil (to be implemented).

Added ClickItemBox, which displays a clickable box with an item in it. Might not be used.

Added test folder, with PlayerDataTest and ItemEnumTest.